validate stream periods on create/update/delete flow events; split HOL validations into separate files for clarity
