+++ packages/subgraph/test/queries/holQueries.ts
@@ -20,7 +20,7 @@ export const getAccount = gql`
         }
     }
 `;
-// TEST-TODO: write the query for getting StreamPeriod here
+
 export const getStream = gql`
     query getStream($id: ID!) {
         response: stream(id: $id) {
@@ -39,6 +39,44 @@ export const getStream = gql`
             flowUpdatedEvents(orderBy: timestamp, orderDirection: asc) {
                 id
             }
+            streamPeriods(orderBy: startedAtTimestamp, orderDirection: asc) {
+                id
+            }
+        }
+    }
+`;
+
+export const getStreamPeriod = gql`
+    query getStreamPeriod($id: ID!) {
+        response: streamPeriod(id: $id) {
+            id
+            flowRate
+
+            token {
+                id
+            }
+            sender {
+                id
+            }
+            receiver {
+                id
+            }
+            startedAtTimestamp
+            startedAtBlockNumber
+            startedAtEvent {
+                id
+            }
+
+            stoppedAtTimestamp
+            stoppedAtBlockNumber
+            stoppedAtEvent {
+                id
+            }
+            stream {
+                id
+            }
+
+            totalAmountStreamed
         }
     }
 `;
